Add an ability to run default IIS website

+++ opsworks_iis/recipes/default.rb
@@ -6,12 +6,19 @@ end
 
 include_recipe "opsworks_iis::urlrewrite"
 
-powershell_script "drop 'Default Web Site' from IIS" do
-  code <<-EOC
-    if ($null -ne (Get-WebSite | where-object { $_.name -eq 'Default Web Site' })) {
-      Remove-Website -Name 'Default Web Site'
-    }
-  EOC
+template ::File.join(node["opsworks_iis"]["doc_root"], "default.htm") do
+  source "default_htm.erb"
+  rights :read, "Everyone"
+end
+
+if node["opsworks_iis"]["remove_default_website"]
+  powershell_script "drop 'Default Web Site' from IIS" do
+    code <<-EOC
+      if ($null -ne (Get-WebSite | where-object { $_.name -eq 'Default Web Site' })) {
+        Remove-Website -Name 'Default Web Site'
+      }
+    EOC
+  end
 end
 
 service "w3svc" do
